Some background check on payment IPN method. Added print request id to the callback URL. Some code organization on controllers package.
