+++ src/main/java/io/github/proxyprint/kitchen/controllers/consumer/PaymentController.java
@@ -1,4 +1,4 @@
-package io.github.proxyprint.kitchen.controllers;
+package io.github.proxyprint.kitchen.controllers.consumer;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
@@ -8,10 +8,14 @@ import com.paypal.base.rest.PayPalRESTException;
 import com.paypal.core.LoggingManager;
 import com.paypal.ipn.IPNMessage;
 import io.github.proxyprint.kitchen.Configuration;
+import io.github.proxyprint.kitchen.models.consumer.Consumer;
+import io.github.proxyprint.kitchen.models.consumer.printrequest.PrintRequest;
 import io.github.proxyprint.kitchen.models.repositories.ConsumerDAO;
 import io.github.proxyprint.kitchen.models.repositories.EmployeeDAO;
+import io.github.proxyprint.kitchen.models.repositories.PrintRequestDAO;
 import io.github.proxyprint.kitchen.models.repositories.UserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
@@ -36,6 +40,8 @@ public class PaymentController {
     @Autowired
     private EmployeeDAO employees;
     @Autowired
+    private PrintRequestDAO printRequests;
+    @Autowired
     private Gson GSON;
 
     @RequestMapping(value = "/paypal/getaccesstoken", method = RequestMethod.GET)
@@ -55,8 +61,9 @@ public class PaymentController {
     }
 
 
-    @RequestMapping(value = "/paypal/ipn", method = RequestMethod.POST)
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    // eCheck - complete
+    @RequestMapping(value = "/paypal/ipn/{printRequestID}", method = RequestMethod.POST)
+    protected void doPost(@PathVariable(value = "printRequestID") long prid, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
         // For a full list of configuration parameters refer in wiki page.
         // (https://github.com/paypal/sdk-core-java/wiki/SDK-Configuration-Parameters)
@@ -66,7 +73,34 @@ public class PaymentController {
         String transactionType = ipnlistener.getTransactionType();
         Map<String,String> map = ipnlistener.getIpnMap();
 
-        LoggingManager.info(this.getClass(), "******* IPN (name:value) pair : "+ map + "  " +
-                "######### TransactionType : "+transactionType+"  ======== IPN verified : "+ isIpnVerified);
+        // Parse content of interest from the IPN notification JSON Body -- :D Todalooo
+        String payerEmail = map.get("payer_email");
+        Double quantity = Double.valueOf(map.get("quantity"));
+        String paymentStatus = map.get("payment_status");
+
+
+        if(payerEmail!=null && quantity!=null && paymentStatus!=null) {
+            Consumer c = consumers.findByEmail(payerEmail);
+            if (c != null) {
+                LoggingManager.info(this.getClass(), "******* IPN (name:value) pair : " + map + "  " +
+                        "######### TransactionType : " + transactionType + "  ======== IPN verified : " + isIpnVerified);
+
+                PrintRequest pr = printRequests.findOne(prid);
+
+                // Divine Condition for secure request background check
+                if(pr!=null && c.getPrintRequests().contains(pr) && c.getEmail().equals(payerEmail)
+                        /*&& pr.getCost()==quantity*/ && paymentStatus.equals(PrintRequest.PAYPAL_COMPLETED_PAYMENT)) {
+                    // The print request may now go to the printshop!
+                    pr.setStatus(PrintRequest.Status.PENDING);
+                    printRequests.save(pr);
+                    return;
+                }
+            } else {
+                LoggingManager.info(this.getClass(), "PayPal transaction ERROR: consumer with email <" + payerEmail + "> is not registred in ProxyPrint.");
+                return;
+            }
+        } else {
+            LoggingManager.info(this.getClass(), "PayPal transaction ERROR: bad IPN JSON body for values payerEmail, quantity, paymentStatus, paymentData");
+        }
     }
 }
